requirements.txt
pandas
beautifulsoup4
yamlloader
mysql-connector-python